+++ libs/causality/src/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from "./lib/CasualAnalysisDashboard/CausalInsightsTab";
+export * from "./lib/CausalAnalysisDashboard/CausalInsightsTab";
